% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findKeyContributors.R
\name{findKeyContributors}
\alias{findKeyContributors}
\title{Identification of Key Contributors}
\usage{
findKeyContributors(relMatrix, metaData, paranIterations=100, paranCentile=99, distMatrix=FALSE, verbose=TRUE)
}
\arguments{
\item{relMatrix}{Pairwise relationship / similarity matrix between samples (N x N)}

\item{metaData}{Data frame for NetView containing ID of individuals ordered as in the similarity matrix (N)}

\item{paranIterations}{Iterations for computing the number of significant components using Horn's Parallel Analysis in Paran [ int, 100 ]}

\item{distMatrix}{Input matrix is a distance matrix; converted to similarity matrix by subtracting from 1 [ bool, FALSE ]}

\item{verbose}{Print computation status and results of Horn's Parallel Analysis [ bool, TRUE ]}

\item{paran.centile}{Significance level (centile used in estimating bias) in Paran [ int, 99 ]}
}
\value{
Data frame containing ordered genetic contribution scores (GCS)
}
\description{
Identification of key contributors using Horn's Parallel Analysis, Eigenvalue Decomposition and
calculation of the Genetic Contribution Score (GCS) as described in Neuditschko et al. (2016).
}

