% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netview.R
\name{netview}
\alias{netview}
\title{NetView R}
\usage{
netview(distMatrix, metaData, k=1:60, cluster=TRUE, ...)
}
\arguments{
\item{distMatrix}{Symmetrical distance matrix for NetView (N x N)}

\item{metaData}{Data frame with meta data for each sample (N), ordered as samples in distMatrix; required columns 'ID', 'Group' and 'Colour'}

\item{tree}{Rooted phylogeny for construction of a cophenetic distance matrix [ phylo, NULL ]}

\item{k}{Range of parameter k for mutual k-nearest-neighbour search [ vector of integers, 10:60 ]}

\item{cluster}{Decorate graph objects with the results of community detection analyses, configure via netviewOptions [ bool, FALSE ]}

\item{mst}{Include edges of the minimum spanning tree associated with the data [ bool, FALSE ]}

\item{networkD3}{Return networks visualizations with networkD3, configure via netviewOptions [ bool, FALSE ]}

\item{save}{Save graphs as GML or visualizations with networkD3 as HTML [ bool, FALSE ]}

\item{project}{Project directory for saving graphs and networkD3 [ str, 'netview' ]}

\item{options}{Options from netviewOptions() [ list, netviewOptions() ]}
}
\value{
List of graph objects or widgets from networkD3.
}
\description{
Generate mutual nearest-neighbour graphs for population analysis and visualization with iGraph and networkD3.
}
\details{
For examples and tutorials, please see our GitHub Repository: \url{https://github.com/esteinig/netview}
}

