% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAdmixture.R
\name{plotAdmixture}
\alias{plotAdmixture}
\title{Admixture Plots}
\usage{
plotAdmixture(qFile="oyster.4.Q", metaData=oysterData, graph=NULL, structurePlot=T, palette="Dark2", colourN=8)
}
\arguments{
\item{qFile}{Output file for ancestry proportions from Admixture (.Q)}

\item{metaData}{Data frame with meta data as for NetView; required columns 'ID', 'Group' and 'Colour'}

\item{graph}{Network object iGraph for generating Admixture Graph [ graph, NULL ]}

\item{structurePlot}{Generate admixture proportion bar plot using Structure Plot [ bool, TRUE ]}

\item{palette}{Colour palette from RColorBrewer or vector of colours (length K) [ str, "Dark2" ]}

\item{colourN}{Number of colours to include in palette, see RColorBrewer [ int, 8 ]}
}
\value{
Graph object with data for plotting admixture proportions or bar plot from Structure Plot.
}
\description{
Generate admixture graphs for NetView or plot admixture proportions with Structure Plot.
}
\details{
For examples and tutorials, please see our GitHub Repository: \url{https://github.com/esteinig/netview}
}

