% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSelection.R
\name{plotSelection}
\alias{plotSelection}
\title{Selection k-Plot}
\usage{
plotSelection(graphs, options=netviewOptions() )
}
\arguments{
\item{graphs}{List of graphs decorated with community-detection results from NetView.}

\item{options}{Options for NetView via netviewOptions [ list, netviewOptons() ]}
}
\value{
List of community data (Data) and plot (Plot).
}
\description{
Generates plot showing the number of detected communities against the parameter k for mutual k-nearest-neighbour graphs from NetView.
}
\details{
For examples and tutorials, please see our GitHub Repository: \url{https://github.com/esteinig/netview}
}

