% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotValidation.R
\name{plotValidation}
\alias{plotValidation}
\title{Admixture Cross-Validation Error}
\usage{
plotValidation(crossError)
}
\arguments{
\item{crossError}{List of cross-validation error values from Admixture, names designate K.}
}
\value{
Cross-validation error plot for output from Admixture.
}
\description{
Convenience function for plotting the cross-validation error output from Admixture.
}
\details{
For examples and tutorials, please see our GitHub Repository: \url{https://github.com/esteinig/netview}
}

