% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAdmixture.R
\name{runAdmixture}
\alias{runAdmixture}
\title{runAdmixture}
\usage{
runAdmixture(filePath="/home/esteinig/admixture/butterflyfish.bed", K=1:10, processors=1, crossValidation=20, plotValidation=FALSE)
}
\arguments{
\item{filePath}{Path to input files, either '.bed', '.ped' or '.geno' (see Admixture Manual)}

\item{K}{Range of K for Admixture [ int vector, 1:10]}

\item{processors}{Number of processors for analysis [ int, 1 ]}

\item{crossValidation}{Number of cross-validation error iterations [ int, 20 ]}

\item{plotValidation}{Return plot of cross-validation error across runs [ bool, FALSE ]}

\item{admixturePath}{Path to admixture executable [ str, NULL ]}
}
\value{
K-named vector containing cross-validation errors or plot of cross-validation error across K.
}
\description{
Convenience function for running Admixture.
}
\details{
Admixture executable can be in $PATH. If this is not working, you can specify the full path to the executable in admixturePath. For examples and tutorials, please see our GitHub Repository: \url{https://github.com/esteinig/netview}
}

