% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAdmixture.R
\name{runAdmixture}
\alias{runAdmixture}
\title{Admixture}
\usage{
runAdmixture("oyster.zip", oysterData, K=1:10, ...)
}
\arguments{
\item{file}{Zipped input data for Admixture containing PLINK binary files, ordered same as metaData for NetView}

\item{project}{Name of temporary output directory and final zipped output file for plotAdmixture() [ str, "admixture_graph" ]}

\item{K}{Range of K for Admixture [ int vector, 1:10]}

\item{processors}{Number of processors for analysis [ int, 1 ]}

\item{crossValidation}{Number of cross-validation error iterations [ int, 20 ]}
}
\value{
Graph object with data for plotting admixture proportions or bar plot from Structure Plot.
}
\description{
Convenience function for running Admixture.
}
\details{
For examples and tutorials, please see our GitHub Repository: \url{https://github.com/esteinig/netview}
\itemize{
     \item{Input files should be generated in PLINK, using the "make-bed" option (.bed, .bim, .fam)}
     \item{Admixture executable needs to be in $PATH}
     \item{If you encounter an error during analysis, please delete the temporary directory and reset your current working directory before running the analysis again}
     }
}

